[refactor] 오늘의 옷장 카드 모달 컴포넌트로 분리 완료
