+++ src/types/Home/index.ts
@@ -1,0 +1,4 @@
+export interface ITodayCardModalProps {
+  isOpen: boolean;
+  setIsOpen: React.Dispatch<React.SetStateAction<boolean>>;
+}
